+++ tests/integration/server-hook/new-middleware/src/routes/page.tsx
@@ -1,0 +1,7 @@
+import { useLoaderData } from '@modern-js/runtime/router';
+
+export default function Page() {
+  const { name } = useLoaderData() as { name: string };
+
+  return <div>Hello {name}</div>;
+}