+++ packages/runtime/plugin-router-v5/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @modern-js/plugin-router-v5
 
+## 2.49.4
+
+### Patch Changes
+
+- @modern-js/plugin@2.49.4
+- @modern-js/runtime-utils@2.49.4
+- @modern-js/types@2.49.4
+- @modern-js/utils@2.49.4
+
 ## 2.49.3
 
 ### Patch Changes